requirements.txt
fastapi
uvicorn
llama-index
openai
sqlalchemy
python-dotenv